% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_dir.R
\name{qa_cache_dir}
\alias{qa_cache_dir}
\title{Get or set the directory used to cache Quantarctica data}
\usage{
qa_cache_dir(path, verbose = FALSE)
}
\arguments{
\item{path}{string: (optional) Values can be "session" (a per-session temporary directory will be used, default), "persistent" (the directory returned by \code{rappdirs::user_cache_dir} will be used), or a string giving the path to the directory to use. An attempt will be made to create the cache directory if it does not exist}

\item{verbose}{logical: show progress messages?}
}
\value{
The path to the cache directory
}
\description{
The cache directory is used to store data locally, so that they can be used offline later. If called with no arguments (i.e. \code{qa_cache_dir()}), this function returns the current cache directory. By default, this is a per-session temporary directory. Calling with a \code{path} argument will set the cache directory to that path.
}
\examples{

## return the current cache directory
qa_cache_dir()

## set the cache directory to a location that persists across
##  user sessions
qa_cache_dir("persistent")

## switch to a per-session cache
qa_cache_dir("session")

}
\seealso{
\code{\link{qa_get}}
}
