% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{qa_dataset}
\alias{qa_dataset}
\title{Retrieve details of a Quantarctica data set}
\usage{
qa_dataset(name, refresh_cache = FALSE, verbose = FALSE)
}
\arguments{
\item{name}{string: the name of the data set}

\item{refresh_cache}{logical: as for \code{qa_get}}

\item{verbose}{logical: show progress messages?}
}
\value{
A tibble
}
\description{
Retrieve details of a Quantarctica data set
}
