% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{qa_datasets}
\alias{qa_datasets}
\title{Available Quantarctica data sets}
\usage{
qa_datasets(cache_directory, refresh_cache = FALSE, verbose = FALSE)
}
\arguments{
\item{cache_directory}{string: as for \code{qa_get}}

\item{refresh_cache}{logical: as for \code{qa_get}}
}
\value{
A tibble with columns \code{id}, \code{name}, \code{source}, and \code{downloaded}
}
\description{
Available Quantarctica data sets
}
\examples{

qa_datasets()

}
\seealso{
\code{\link{qa_get}}
}
