% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{qa_find_shapefile}
\alias{qa_find_shapefile}
\title{Find the shapefiles amongst a set of files}
\usage{
qa_find_shapefile(result)
}
\arguments{
\item{result}{: as returned by \code{qa_get}}
}
\value{
A character vector of paths to shapefiles
}
\description{
Find the shapefiles amongst a set of files
}
