% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{qa_get}
\alias{qa_get}
\title{Fetch a Quantarctica data set}
\usage{
qa_get(dataset, cache_directory = qa_cache_dir(), refresh_cache = 0,
  verbose = FALSE, shapefile_reader, raster_reader)
}
\arguments{
\item{dataset}{string or tibble: the name of the data set or a tibble as returned by \code{qa_dataset}}

\item{cache_directory}{string: the cache directory to use. As for the \code{path} parameter to the \code{\link{qa_cache_dir}} function}

\item{refresh_cache}{numeric: 0 = do not overwrite existing files, 1 = overwrite if the remote file is newer than the local copy, 2 = always overwrite existing files}

\item{verbose}{logical: show progress messages?}

\item{shapefile_reader}{function: function to use to read shapefiles. By default this is \code{raster::shapefile}}

\item{raster_reader}{function: function to use to read raster files (TIFF, JP2, VRT). By default this is \code{raster::raster}}
}
\value{
TBD
}
\description{
Fetch a Quantarctica data set
}
