% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_dir.R
\name{qa_is_cached}
\alias{qa_is_cached}
\title{Has a dataset been downloaded to the local cache?}
\usage{
qa_is_cached(dataset, cache_directory = qa_cache_dir(), verbose = FALSE)
}
\arguments{
\item{dataset}{string or qa_dataset: the dataset object, or name of the dataset}

\item{cache_directory}{string: the cache directory to check. As for the \code{path} parameter to the \code{\link{qa_cache_dir}} function}

\item{verbose}{logical: show progress messages?}
}
\value{
Logical
}
\description{
Has a dataset been downloaded to the local cache?
}
\examples{
\dontrun{
  ds <- qa_dataset("ADD Simple basemap")
  qa_is_cached(ds) ## FALSE
  x <- qa_get(ds)
  qa_is_cached(ds) ## TRUE
}

}
\seealso{
\code{\link{qa_dataset}}, \code{\link{qa_cache_dir}}
}
