% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrors.R
\name{qa_mirrors}
\alias{qa_mirrors}
\title{A list of Quantarctic mirror sites}
\usage{
qa_mirrors()
}
\value{
A tibble with columns \code{country}, \code{protocol}, and \code{url}
}
\description{
Choose the mirror closest to you for the fastest and most reliable download.
}
\examples{

qa_mirrors()

}
\references{
http://quantarctica.npolar.no/downloads/
}
