% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{qa_get}
\alias{qa_get}
\title{Fetch a Quantarctica data set}
\usage{
qa_get(dataset, cache_directory, refresh_cache = FALSE,
  verbose = FALSE)
}
\arguments{
\item{dataset}{string: the name of the data set}

\item{cache_directory}{string: (optional) cache the data locally in this directory, so that they can be used offline later. Values can be "session" (a per-session temporary directory will be used, default), "persistent" (the directory returned by \code{rappdirs::user_cache_dir} will be used), or a string giving the path to the directory to use. Use \code{NULL} for no caching. An attempt will be made to create the cache directory if it does not exist. Use \code{refresh_cache = TRUE} to refresh the cached data if necessary}

\item{refresh_cache}{logical: if TRUE, and data already exist in the cache_directory, they will be refreshed. If FALSE, the cached data will be used}

\item{verbose}{logical: show progress messages?}
}
\value{
TBD
}
\description{
Fetch a Quantarctica data set
}
