% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantarcticR-package.R
\docType{package}
\name{quantarcticR-package}
\alias{quantarcticR}
\alias{quantarcticR-package}
\title{quantarcticR: Access to Quantarctica Data Sets}
\description{
Quantarctica is a collection of Antarctic geographical datasets which works with the free, cross-platform, open-source software QGIS. It includes community-contributed, peer-reviewed data from ten different scientific themes and a professionally-designed basemap. This package provides access to Quantarctica data sets.
}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

Authors:
\itemize{
  \item Kim Fitter \email{kim@validly.co}
  \item Charlène Guillaumot
  \item David Ranzolin \email{daranzolin@gmail.com}
  \item Marc Desgroseilliers
}

}
\keyword{internal}
